requirements.txt
streamlit
pandas
openai
langchain
faiss-cpu
python-dotenv
langchain-openai